test modularization cleanup
