+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/table/TableAggregationTest.scala
@@ -1,12 +1,11 @@
-package pl.touk.nussknacker.devmodel
+package pl.touk.nussknacker.engine.flink.table
 
 import com.typesafe.config.ConfigFactory
 import org.scalatest.Inside
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
-import pl.touk.nussknacker.devmodel.TableAggregationTest.TestRecord
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
-import pl.touk.nussknacker.engine.flink.table.FlinkTableComponentProvider
+import pl.touk.nussknacker.engine.flink.table.TableAggregationTest.TestRecord
 import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregationFactory
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.util.test.TestScenarioRunner
